+++ lib/ts/fetch.ts
@@ -16,7 +16,7 @@ import { PROCESS_STATE, ProcessState } from "./processState";
 import { supported_fdi } from "./version";
 import Lock from "browser-tabs-lock";
 import { shouldDoInterceptionBasedOnUrl } from "./utils";
-import { RecipeInterface, NormalisedInputType } from "./types";
+import { RecipeInterface, NormalisedInputType, ClaimValidationError } from "./types";
 import CookieHandlerReference from "./utils/cookieHandler";
 import WindowHandlerReference from "./utils/windowHandler";
 import { logDebugMessage } from "./logger";
@@ -298,8 +298,8 @@ export default class AuthHttpRequest {
                     }
                     logDebugMessage("doRequest: Retrying original request");
                 } else {
-                    if (response.status === AuthHttpRequest.config.missingClaimStatusCode) {
-                        onMissingClaimResponse(response.headers.get("missing-claim-id")!);
+                    if (response.status === AuthHttpRequest.config.invalidClaimStatusCode) {
+                        onInvalidClaimResponse(JSON.parse(response.headers.get("invalid-claim")!));
                     }
                     const antiCsrfToken = response.headers.get("anti-csrf");
                     if (antiCsrfToken) {
@@ -538,10 +538,10 @@ export function onTokenUpdate() {
     });
 }
 
-export function onMissingClaimResponse(claimId: string) {
+export function onInvalidClaimResponse(claimValidationError: ClaimValidationError) {
     AuthHttpRequest.config.onHandleEvent({
         action: "API_INVALID_CLAIM",
-        claimId
+        claimValidationError
     });
 }